Refactor: Extract Firestore operations into services module
- Created src/services/firestore.js with three focused functions:
  - loadLinksFromFirestore(userId)
  - saveLinkToFirestore(userId, url, title)
  - deleteLinkFromFirestore(userId, linkId)

- Simplified popup.js from 439 to 402 lines (37 lines removed)
- Removed Firestore imports and db dependency from popup.js
- Improved code maintainability and reusability
- Service functions now ready for website dashboard integration
